# backend/api/gen/go/admin/service/v1/i_payee_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_payee.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-ledger/api/gen/go/ledger/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationPayeeServiceCreate = "/admin.service.v1.PayeeService/Create"
const OperationPayeeServiceDelete = "/admin.service.v1.PayeeService/Delete"
const OperationPayeeServiceGet = "/admin.service.v1.PayeeService/Get"
const OperationPayeeServiceList = "/admin.service.v1.PayeeService/List"
const OperationPayeeServiceListAll = "/admin.service.v1.PayeeService/ListAll"
const OperationPayeeServiceToggle = "/admin.service.v1.PayeeService/Toggle"
const OperationPayeeServiceToggleCanExpense = "/admin.service.v1.PayeeService/ToggleCanExpense"
const OperationPayeeServiceToggleCanIncome = "/admin.service.v1.PayeeService/ToggleCanIncome"
const OperationPayeeServiceUpdate = "/admin.service.v1.PayeeService/Update"

type PayeeServiceHTTPServer interface {
	Create(context.Context, *v11.CreatePayeeRequest) (*v11.Payee, error)
	Delete(context.Context, *v11.DeletePayeeRequest) (*emptypb.Empty, error)
	Get(context.Context, *v11.GetPayeeRequest) (*v11.Payee, error)
	List(context.Context, *v1.PagingRequest) (*v11.ListPayeeResponse, error)
	ListAll(context.Context, *v11.ListAllPayeeRequest) (*v11.ListPayeeResponse, error)
	Toggle(context.Context, *v11.TogglePayeeRequest) (*v11.Payee, error)
	ToggleCanExpense(context.Context, *v11.TogglePayeeRequest) (*v11.Payee, error)
	ToggleCanIncome(context.Context, *v11.TogglePayeeRequest) (*v11.Payee, error)
	Update(context.Context, *v11.UpdatePayeeRequest) (*v11.Payee, error)
}

func RegisterPayeeServiceHTTPServer(s *http.Server, srv PayeeServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/payees", _PayeeService_List20_HTTP_Handler(srv))
	r.GET("/admin/v1/payees-all", _PayeeService_ListAll5_HTTP_Handler(srv))
	r.GET("/admin/v1/payees/{id}", _PayeeService_Get19_HTTP_Handler(srv))
	r.POST("/admin/v1/payees", _PayeeService_Create14_HTTP_Handler(srv))
	r.PUT("/admin/v1/payees/{id}", _PayeeService_Update14_HTTP_Handler(srv))
	r.DELETE("/admin/v1/payees/{id}", _PayeeService_Delete15_HTTP_Handler(srv))
	r.PATCH("/admin/v1/payees/{id}/toggle", _PayeeService_Toggle2_HTTP_Handler(srv))
	r.PATCH("/admin/v1/payees/{id}/toggle-can-expense", _PayeeService_ToggleCanExpense1_HTTP_Handler(srv))
	r.PATCH("/admin/v1/payees/{id}/toggle-can-income", _PayeeService_ToggleCanIncome1_HTTP_Handler(srv))
}

func _PayeeService_List20_HTTP_Handler(srv PayeeServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayeeServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListPayeeResponse)
		return ctx.Result(200, reply)
	}
}

func _PayeeService_ListAll5_HTTP_Handler(srv PayeeServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.ListAllPayeeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayeeServiceListAll)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListAll(ctx, req.(*v11.ListAllPayeeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListPayeeResponse)
		return ctx.Result(200, reply)
	}
}

func _PayeeService_Get19_HTTP_Handler(srv PayeeServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetPayeeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayeeServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetPayeeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Payee)
		return ctx.Result(200, reply)
	}
}

func _PayeeService_Create14_HTTP_Handler(srv PayeeServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreatePayeeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayeeServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreatePayeeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Payee)
		return ctx.Result(200, reply)
	}
}

func _PayeeService_Update14_HTTP_Handler(srv PayeeServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdatePayeeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayeeServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdatePayeeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Payee)
		return ctx.Result(200, reply)
	}
}

func _PayeeService_Delete15_HTTP_Handler(srv PayeeServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeletePayeeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayeeServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeletePayeeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _PayeeService_Toggle2_HTTP_Handler(srv PayeeServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.TogglePayeeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayeeServiceToggle)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Toggle(ctx, req.(*v11.TogglePayeeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Payee)
		return ctx.Result(200, reply)
	}
}

func _PayeeService_ToggleCanExpense1_HTTP_Handler(srv PayeeServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.TogglePayeeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayeeServiceToggleCanExpense)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ToggleCanExpense(ctx, req.(*v11.TogglePayeeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Payee)
		return ctx.Result(200, reply)
	}
}

func _PayeeService_ToggleCanIncome1_HTTP_Handler(srv PayeeServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.TogglePayeeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayeeServiceToggleCanIncome)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ToggleCanIncome(ctx, req.(*v11.TogglePayeeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Payee)
		return ctx.Result(200, reply)
	}
}

type PayeeServiceHTTPClient interface {
	Create(ctx context.Context, req *v11.CreatePayeeRequest, opts ...http.CallOption) (rsp *v11.Payee, err error)
	Delete(ctx context.Context, req *v11.DeletePayeeRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	Get(ctx context.Context, req *v11.GetPayeeRequest, opts ...http.CallOption) (rsp *v11.Payee, err error)
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListPayeeResponse, err error)
	ListAll(ctx context.Context, req *v11.ListAllPayeeRequest, opts ...http.CallOption) (rsp *v11.ListPayeeResponse, err error)
	Toggle(ctx context.Context, req *v11.TogglePayeeRequest, opts ...http.CallOption) (rsp *v11.Payee, err error)
	ToggleCanExpense(ctx context.Context, req *v11.TogglePayeeRequest, opts ...http.CallOption) (rsp *v11.Payee, err error)
	ToggleCanIncome(ctx context.Context, req *v11.TogglePayeeRequest, opts ...http.CallOption) (rsp *v11.Payee, err error)
	Update(ctx context.Context, req *v11.UpdatePayeeRequest, opts ...http.CallOption) (rsp *v11.Payee, err error)
}

type PayeeServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewPayeeServiceHTTPClient(client *http.Client) PayeeServiceHTTPClient {
	return &PayeeServiceHTTPClientImpl{client}
}

func (c *PayeeServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreatePayeeRequest, opts ...http.CallOption) (*v11.Payee, error) {
	var out v11.Payee
	pattern := "/admin/v1/payees"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPayeeServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PayeeServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeletePayeeRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/payees/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPayeeServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PayeeServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetPayeeRequest, opts ...http.CallOption) (*v11.Payee, error) {
	var out v11.Payee
	pattern := "/admin/v1/payees/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPayeeServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PayeeServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListPayeeResponse, error) {
	var out v11.ListPayeeResponse
	pattern := "/admin/v1/payees"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPayeeServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PayeeServiceHTTPClientImpl) ListAll(ctx context.Context, in *v11.ListAllPayeeRequest, opts ...http.CallOption) (*v11.ListPayeeResponse, error) {
	var out v11.ListPayeeResponse
	pattern := "/admin/v1/payees-all"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPayeeServiceListAll))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PayeeServiceHTTPClientImpl) Toggle(ctx context.Context, in *v11.TogglePayeeRequest, opts ...http.CallOption) (*v11.Payee, error) {
	var out v11.Payee
	pattern := "/admin/v1/payees/{id}/toggle"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPayeeServiceToggle))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PayeeServiceHTTPClientImpl) ToggleCanExpense(ctx context.Context, in *v11.TogglePayeeRequest, opts ...http.CallOption) (*v11.Payee, error) {
	var out v11.Payee
	pattern := "/admin/v1/payees/{id}/toggle-can-expense"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPayeeServiceToggleCanExpense))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PayeeServiceHTTPClientImpl) ToggleCanIncome(ctx context.Context, in *v11.TogglePayeeRequest, opts ...http.CallOption) (*v11.Payee, error) {
	var out v11.Payee
	pattern := "/admin/v1/payees/{id}/toggle-can-income"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPayeeServiceToggleCanIncome))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PayeeServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdatePayeeRequest, opts ...http.CallOption) (*v11.Payee, error) {
	var out v11.Payee
	pattern := "/admin/v1/payees/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPayeeServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}